test(api): factor out test for attribute completeness
move them to its own file
