+++ tests/integrationtests/api/test_indicators.py
@@ -6,7 +6,6 @@ Validate the response from requests to the `/indicators` endpoint of the API.
 import pytest
 from schema import Optional, Or, Schema
 
-from ohsome_quality_api.attributes.definitions import get_attributes
 from tests.integrationtests.utils import oqapi_vcr
 
 ENDPOINT = "/indicators/"
@@ -119,67 +118,6 @@ def test_indicators(
     }
     response = client.post(endpoint, json=parameters, headers=headers)
     assert schema.is_valid(response.json())
-
-
-@oqapi_vcr.use_cassette
-def test_indicators_attribute_completeness(
-    client,
-    bpolys,
-    headers,
-    schema,
-):
-    endpoint = ENDPOINT + "attribute-completeness"
-    parameters = {"bpolys": bpolys, "topic": "building-count", "attributes": ["height"]}
-    response = client.post(endpoint, json=parameters, headers=headers)
-    assert schema.is_valid(response.json())
-
-
-def test_indicators_attribute_completeness_without_attribute(
-    client,
-    bpolys,
-    headers,
-    schema,
-):
-    endpoint = ENDPOINT + "attribute-completeness"
-    parameters = {
-        "bpolys": bpolys,
-        "topic": "building-count",
-    }
-    response = client.post(endpoint, json=parameters, headers=headers)
-    assert response.status_code == 422
-    content = response.json()
-    assert content["type"] == "RequestValidationError"
-
-
-def test_indicators_attribute_completeness_with_invalid_attribute_for_topic(
-    client,
-    bpolys,
-    headers,
-    schema,
-):
-    endpoint = ENDPOINT + "attribute-completeness"
-    parameters = {
-        "bpolys": bpolys,
-        "topic": "building-count",
-        # the following attribute is not valid for topic 'building-count'
-        "attributes": ["maxspeed"],
-    }
-    response = client.post(endpoint, json=parameters, headers=headers)
-    assert response.status_code == 422
-    content = response.json()
-
-    message = content["detail"][0]["msg"]
-    all_attributes_for_topic = [
-        attribute for attribute in (get_attributes()["building-count"])
-    ]
-
-    expected = (
-        "Invalid combination of attribute and topic: maxspeed and building-count. "
-        "Topic 'building-count' supports these attributes: {}"
-    ).format(all_attributes_for_topic)
-
-    assert message == expected
-    assert content["type"] == "AttributeTopicCombinationError"
 
 
 @oqapi_vcr.use_cassette
