+++ tests/integrationtests/approved/indicators/test_attribute_completeness.py::TestCalculation::test_calculate[indicator0]..approved.txt
@@ -1,0 +1,1 @@
+28.8% of all "building count" features (all: 29936 elements) in your area of interest have the selected additional attributes height of buildings (matched: 8630 elements).
