+++ tests/integrationtests/approved/indicators/test_attribute_completeness.py::TestCalculation::test_calculate[indicator0].approved.txt
@@ -1,0 +1,1 @@
+39.8% of all "building count" features (all: 30263 elements) in your area of interest have the selected additional attributes height of buildings (matched: 12059 elements).
