+++ tests/integrationtests/approved/indicators/test_attribute_completeness.py::TestCalculation::test_calculate[indicator1].approved.txt
@@ -1,0 +1,1 @@
+59.7% of all "building count" features (all: 30237 elements) in your area of interest have the selected additional attributes Height (matched: 18057 elements).
