+++ tests/integrationtests/approved/indicators/test_attribute_completeness.py::test_create_description_attribute_keys_multiple.approved.txt
