+++ tests/integrationtests/approved/indicators/test_attribute_completeness.py::test_create_description_attribute_keys_single.approved.txt
@@ -1,0 +1,1 @@
+20.0% of all "building count" features (all: 10 elements) in your area of interest have the selected additional attributes height of buildings (matched: 2 elements).
