--- tests/integrationtests/fixtures/approved/test_attribute_completeness-TestFigure-test_create_figure-indicator0 (setup).approved.txt
@@ -1,1 +1,0 @@
-28.8% of all "building count" features (all: 29936 elements) in your area of interest have the selected additional attribute height of buildings (matched: 8630 elements).
