--- tests/integrationtests/fixtures/approved/test_attribute_completeness-TestFigure-test_create_figure-indicator1 (setup).approved.txt
@@ -1,1 +1,0 @@
-59.7% of all "building count" features (all: 30237 elements) in your area of interest have the selected additional attribute Height (matched: 18057 elements).
