--- tests/integrationtests/fixtures/approved/test_attribute_completeness-test_create_description_attribute_keys_single.approved.txt
@@ -1,1 +1,0 @@
-20.0% of all "building count" features (all: 10 elements) in your area of interest have the selected additional attribute height of buildings (matched: 2 elements).
